refactor address and availability section
